refactor: изменить формат рисунков на jpg и добавить ссылку на gh-pages
